+++ src/vscode/treeviews/Connections/FabricConnectionsTreeProvider.ts
@@ -52,6 +52,13 @@ export class FabricConnectionsTreeProvider implements vscode.TreeDataProvider<Fa
 		if(tree_item && tree_item.collapsibleState == vscode.TreeItemCollapsibleState.None) {
 			tree_item = tree_item.parent;
 		}
+		if (tree_item) {
+			tree_item.refreshApiPaths.forEach(path => FabricApiService.clearCache(path));
+		}
+		else {
+			FabricApiService.clearCache("/v1/gateways");
+			FabricApiService.clearCache("/v1/connections");
+		}
 		this._onDidChangeTreeData.fire(tree_item);
 	}
 
@@ -106,7 +113,10 @@ export class FabricConnectionsTreeProvider implements vscode.TreeDataProvider<Fa
 						}
 					}
 
-					const gateway = item.gatewayId ?? item.connectivityType;
+					let gateway = item.gatewayId;
+					if(["PersonalCloud", "ShareableCloud", "OnPremisesGatewayPersonal"].includes(item.connectivityType)) {
+						gateway = item.connectivityType;
+					}
 
 					// for ShareableCloud connectsions, the gatewayId is not set, so we need to use the connectivityType to group them
 					if (!gateways.has(gateway)) {